Adding the MarkAsRead functionality

+++ DatingApp-SPA/src/app/members/photo-editor/photo-editor.component.ts
@@ -9,7 +9,7 @@ import { AuthService } from 'src/app/_services/auth.service';
 @Component({
   selector: 'app-photo-editor',
   templateUrl: './photo-editor.component.html',
-  styleUrls: ['./photo-editor.component.css']
+  styleUrls: ['./photo-editor.component.css'],
 })
 export class PhotoEditorComponent implements OnInit {
   @Input() photos: Photo[];
@@ -45,10 +45,10 @@ export class PhotoEditorComponent implements OnInit {
       allowedFileType: ['image'],
       removeAfterUpload: true,
       autoUpload: false,
-      maxFileSize: 10 * 1024 * 1024
+      maxFileSize: 10 * 1024 * 1024,
     });
 
-    this.uploader.onAfterAddingFile = file => {
+    this.uploader.onAfterAddingFile = (file) => {
       file.withCredentials = false;
     };
 
@@ -60,7 +60,7 @@ export class PhotoEditorComponent implements OnInit {
           url: res.url,
           dateAdded: res.dateAdded,
           description: res.description,
-          isMain: res.isMain
+          isMain: res.isMain,
         };
         this.photos.push(photo);
         if (photo.isMain) {
@@ -80,7 +80,7 @@ export class PhotoEditorComponent implements OnInit {
       .setMainPhoto(this.authService.decodedToken.nameid, photo.id)
       .subscribe(
         () => {
-          this.currentMain = this.photos.filter(p => p.isMain === true)[0];
+          this.currentMain = this.photos.filter((p) => p.isMain === true)[0];
           this.currentMain.isMain = false;
           photo.isMain = true;
           this.authService.changeMemberPhoto(photo.url);
@@ -90,24 +90,24 @@ export class PhotoEditorComponent implements OnInit {
             JSON.stringify(this.authService.currentUser)
           );
         },
-        error => {
+        (error) => {
           this.alertify.error(error);
         }
       );
   }
   deletePhoto(id: number) {
-    this.alertify.conform('Are you sure?', () => {
+    this.alertify.confirm('Are you sure?', () => {
       this.userService
         .deletePhoto(this.authService.decodedToken.nameid, id)
         .subscribe(
           () => {
             this.photos.splice(
-              this.photos.findIndex(p => p.id === id),
+              this.photos.findIndex((p) => p.id === id),
               1
             );
             this.alertify.success('Photo has been deleted');
           },
-          error => {
+          (error) => {
             this.alertify.error('Failed to delete the photo');
           }
         );
